added 'context' class
The goal of context is to clean up code by reducing the variables sent to scripts.

Rather than sending 8-15 variables to a script, send it the entire context of dandere2x and the variables it needs.

This reduces overall lines of code needed in dandere2x.py and makes adding additions in the future easier.